+++ src/draw_cmd_header.rs
@@ -1,12 +1,12 @@
-use libc::c_uint;
+use libc::{c_uint, size_t};
 use crate::type_defs::ImTextureID;
 use crate::vec4::ImVec4;
 
 // [Internal] For use by ImDrawList
-#[derive(Default,Debug,Clone)]
+#[derive(Default,Debug,Clone,Copy)]
 pub struct ImDrawCmdHeader
 {
 pub ClipRect:  ImVec4,
 pub TextureId:  ImTextureID,
-pub VtxOffset:  c_uint,
+pub VtxOffset:  size_t,
 }
